time function first draft
